requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosqlite==0.19.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
requests==2.31.0
pytz==2023.3
schedule==1.2.0
cryptography==41.0.7
redis==5.0.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
pillow==10.1.0
